Enhance wrapping with parameterless ctor
When wrapping Il2Cpp objects into injected types that have only parameterless ctor, always call the ctor. Because the ctor can be user-defined (e.g. it calls DerivedConstructorPointer), we add ability to mark object as wrapped in which case the object returned by DerivedConstructorPointer is ignored (and GC'd by Il2Cpp). This allows to always call the parameterless ctor on injected types.