Fix newobj Il2Cpp*Array<T>.ctor reference caching
Because the generic type parameter can change multiple times within the same assembly, we cannot make the reference just Lazy (otherwise it returns the same TypeReference for each attempt to assign the generic parameter). To sole this, we add a simple Memoize wrapper which allows to cache the reference by the generic type parameter.