Find GarbageCollector::RunFinalizer with xref scan as fallback
The current `RunFinalizer` signatures can not reliably matches address of `RunFinalizer`, so fallback to xref scan as a last resort.

`RunFinalizer` is used in `RegisterFinalizer`, and `RegisterFinalizer` is used in icall GC._ReRegisterForFinalize which is relatively simple compared to other users of `RegisterFinalizer`.
So we can traversal into `GC._ReRegisterForFinalize` to find address of `RunFinalizer`.

The xref scan handles both inlined and not inlined cases.